Move config helpers to a more useful place
